+++ docu/src/main/java/io/github/factoryfx/docu/initializr/ServerMain.java
@@ -1,4 +1,4 @@
-package io.github.factoryfx.docu.starter;
+package io.github.factoryfx.docu.initializr;
 
 import java.lang.String;