+++ CHANGELOG.md
@@ -15,7 +15,7 @@
   
 * **builder:** new builder API for microservice setup
 
-* **starter:** One time java code generator for initial project setup to simplify the first steps with factoryfx.
+* **initializr:** One time java code generator for initial project setup to simplify the first steps with factoryfx.
 
 * **testing:** Added the ability to set a mock liveobject for a factory.